no-issue: move nextflow_params_handler and test to utils, logger etc to validation utils, should be just structural changes
