+++ modules/validation/validation-pkg/validation_pkg/__init__.py
@@ -19,7 +19,7 @@ from validation_pkg.validators.feature_validator import FeatureValidator
 from validation_pkg.validators.feature_validator import FeatureOutputMetadata
 from validation_pkg.validators.interfile_read import ReadXReadSettings, readxread_validation
 from validation_pkg.validators.interfile_genome import GenomeXGenomeSettings, genomexgenome_validation
-from validation_pkg.logger import setup_logging, get_logger
+from validation_pkg.utils.logger import setup_logging, get_logger
 from validation_pkg.report import ValidationReport
 
 # Functional API imports
