+++ modules/validation/validation-pkg/validation_pkg/utils/file_handler.py
@@ -105,7 +105,7 @@ def _log_compression_tool(tool_name: str, threads: int, is_parallel: bool, insta
         return
 
     try:
-        from validation_pkg.logger import get_logger
+        from validation_pkg.utils.logger import get_logger
         logger = get_logger()
 
         if is_parallel:
