+++ modules/validation/validation-pkg/validation_pkg/validators/interfile_read.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from typing import List, Dict, Optional, Any
 from ..utils.base_settings import BaseSettings
 from ..exceptions import ReadValidationError
-from ..logger import get_logger
+from ..utils.logger import get_logger
 
 
 @dataclass
